+++ src/application/use-cases/CreateCinema/CreateCinemaUseCase.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Inject, Injectable } from "@nestjs/common";
 import type { ICinemaRepository } from "../../../domain/repositories";
 import { Cinema } from "../../../domain/entities/Cinema";
 import { CreateCinemaInput, CreateCinemaOutput } from "./CreateCinemaDTO";
@@ -6,7 +6,10 @@ import { CreateCinemaValidator } from "./CreateCinemaValidator";
 
 @Injectable()
 export class CreateCinemaUseCase {
-  constructor(private cinemaRepository: ICinemaRepository) {}
+  constructor(
+    @Inject('ICinemaRepository')
+    private readonly cinemaRepository: ICinemaRepository
+) {}
 
   async execute(input: CreateCinemaInput): Promise<CreateCinemaOutput> {
     const errors = CreateCinemaValidator.validate(input);