+++ src/modules/movies/movies.module.ts
@@ -1,4 +1,18 @@
 import { Module } from '@nestjs/common';
+import { PrismaService } from '../../prisma/prisma.service';
+import { PrismaMovieRepository } from '../../infrastructure/database/repositories/PrismaMovieRepository';
+import { CreateMovieUseCase } from '../../application/use-cases/CreateMovie/CreateMovieUseCase';
 
-@Module({})
+@Module({
+  providers: [
+    PrismaService,
+    PrismaMovieRepository,
+    CreateMovieUseCase,
+    {
+      provide: 'IMovieRepository',
+      useClass: PrismaMovieRepository,
+    },
+  ],
+  exports: [CreateMovieUseCase],
+})
 export class MoviesModule {}